Deep-Object-Damage-Analysis-main/requirements.txt
argparse
pillow
torch
opencv-python-headless
numpy
tensorflow
flask
werkzeug
requests
ultralytics==8.0.55
scikit-image
scikit-learn